PlayerPanel: Support two lines of text and Updating Instruction Panel
Adds support for 2 lines of custom message
text.

New Instruction Panel

Instruction Panel finished

Test Fade In/Out

Debug Message

We needed to update the instruction panel to see to the patient what is
the good position to have in order to play the game. So the instruction
panel is now updating but we also need to make a fade out to the
calibration panel. However, we will wait the fade In/Out script manager
before doing this.