Data Saving System
We needed to update UI during a save to prevent a save being incomplete.
So for this, we have a warning message panel which appeared when we try
to quit the app during the save. Moreover, we have an indicator of the
total files that we need to save and how many are saved. When the save
is done, we can restart a game.